+++ ptolemy-py/python/ptolemy_client/v1/client.py
@@ -1,35 +1,13 @@
 """Ptolemy Client."""
 
-from typing import Dict, Any, Optional, Self, List, TypeVar, Generic
-from enum import StrEnum
+from typing import Dict, Any, Optional, List
 from uuid import UUID, uuid4
 from pydantic import BaseModel, Field, PrivateAttr
 
+from .tier import Tier
+from .io import IO
+
 Parameters = Dict[str, Any]
-
-T = TypeVar("T")
-
-class Tier(StrEnum):
-    """Tier."""
-
-    SYSTEM = "system"
-    SUBSYSTEM = "subsystem"
-    COMPONENT = "component"
-    SUBCOMPONENT = "subcomponent"
-
-    def child(self) -> Self:
-        """Get child tier."""
-
-        if self == Tier.SYSTEM:
-            return Tier.SUBSYSTEM
-
-        if self == Tier.SUBSYSTEM:
-            return Tier.COMPONENT
-
-        if self == Tier.COMPONENT:
-            return Tier.SUBCOMPONENT
-
-        raise ValueError("Cannot create a child trace of a subcomponent.")
 
 class Ptolemy(BaseModel):
     """Ptolemy Client."""
@@ -46,14 +24,6 @@ class Ptolemy(BaseModel):
             raise ValueError("Workspace ID must be set.")
 
         return self._workspace_id
-
-class IO(BaseModel, Generic[T]):
-    """IO object."""
-
-    parent_id: UUID
-    id_: UUID = Field(default_factory=uuid4, alias="id")
-    field_name: str
-    field_value: T
 
 class Trace(BaseModel):
     """Trace."""
